testing/requirements.txt
aniso8601==9.0.1
apispec==6.4.0
blinker==1.7.0
click==8.1.7
flask==3.0.0
Flask-API==3.1
flask-rest-api==0.17.1
Flask-RESTful==0.3.10
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
marshmallow==3.20.2
packaging==23.2
pytz==2023.3.post1
six==1.16.0
webargs==8.4.0
werkzeug==3.0.1
zipp==3.17.0
pyyaml
